Order

+++ Project/Loothub/src/main/java/com/company/springboot/dao/IOrdersDao.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface IOrdersDao {
     public List<Orders> listAll();
     public Orders save(Orders orders);
-    public Orders get(Long id);
-    public void delete(Long id);
+    public Orders get(int id);
+    public void delete(int id);
     public void update(Orders orders);
     public Orders findByUserId(Integer userId);
 }
