+++ Project/Loothub/src/main/java/com/company/springboot/services/IOrdersServices.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface IOrdersServices {
     public List<Orders> listAll();
     public Orders save(Orders orders);
-    public Orders get(Long id);
-    public void delete(Long id);
+    public Orders get(int id);
+    public void delete(int id);
     public void update(Orders orders);
-    public Orders findByUserId(Integer userId);
+    public Orders findByUserId(int userId);
     
 }
